+++ frame/interface/AFILogicThreadManager.h
@@ -1,0 +1,41 @@
+/*
+* This source file is part of ARK
+* For the latest info, see https://github.com/QuadHex
+*
+* Copyright (c) 2013-2018 QuadHex authors.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*
+*/
+
+#ifndef _AFITHREADMANAGER_H
+#define _AFITHREADMANAGER_H
+
+
+namespace ark
+{
+
+    class AFILogicThreadManager
+    {
+    public:
+        AFILogicThreadManager() {};
+        virtual ~AFILogicThreadManager() {};
+
+        virtual void CheckThreadList() = 0;
+
+        virtual int64_t GetMainThreadChekInterval() = 0;
+    };
+
+}
+
+#endif //_AFITHREADMANAGER_H
